optional stats and dbs_portal

+++ applications/portal/core/libraries/stats.php
@@ -7,6 +7,7 @@ class Stats {
 	
 	private $CI;
 	private $db;
+	private $useStats;
 
 	/**
 	 * Register a page view and essential information
@@ -17,37 +18,40 @@ class Stats {
 	 */
 	public function registerPageView($registry_object_id=null)
 	{
-		$values = array();
-
-		// The server time of the request
-		$values['timestamp'] = time();
-
-		// The request URI (what URI/page was called)
-		if(isset($_SERVER['REQUEST_URI']) && $request_uri = $_SERVER['REQUEST_URI'])
+		if($this->collectStats)
 		{
-			$values['request_uri'] = $request_uri;
+			$values = array();
+
+			// The server time of the request
+			$values['timestamp'] = time();
+
+			// The request URI (what URI/page was called)
+			if(isset($_SERVER['REQUEST_URI']) && $request_uri = $_SERVER['REQUEST_URI'])
+			{
+				$values['request_uri'] = $request_uri;
+			}
+
+			// The page which referred us here (if any)
+			if(isset($_SERVER['HTTP_REFERER']) && $referer = $_SERVER['HTTP_REFERER'])
+			{
+				$values['referer'] = $referer;
+			}
+
+			// Details about the user that browsed here
+			$values['ip_address'] = $this->CI->input->ip_address();
+			$values['user_agent'] = $this->CI->input->user_agent();
+
+			// Including their login id if they are already logged in
+			if($this->CI->user->loggedIn())
+			{
+				$values['login_identifier'] = $this->CI->user->identifier();
+			}
+
+			// Optionally, if this page is a view of a registry object
+			if ($registry_object_id) { $values['registry_object_id'] = $registry_object_id; }
+
+			$this->db->insert('page_views', $values);
 		}
-
-		// The page which referred us here (if any)
-		if(isset($_SERVER['HTTP_REFERER']) && $referer = $_SERVER['HTTP_REFERER'])
-		{
-			$values['referer'] = $referer;
-		}
-
-		// Details about the user that browsed here
-		$values['ip_address'] = $this->CI->input->ip_address();
-		$values['user_agent'] = $this->CI->input->user_agent();
-
-		// Including their login id if they are already logged in
-		if($this->CI->user->loggedIn())
-		{
-			$values['login_identifier'] = $this->CI->user->identifier();
-		}
-
-		// Optionally, if this page is a view of a registry object
-		if ($registry_object_id) { $values['registry_object_id'] = $registry_object_id; }
-
-		$this->db->insert('page_views', $values);
 	}
 
 	/**
@@ -59,29 +63,32 @@ class Stats {
 	 */
 	public function registerClick($source_url, $target_url, $note=null)
 	{
-		$values = array();
-
-		// The server time of the request
-		$values['timestamp'] = time();
-
-		// The source URI (what page was the link originally on)
-		$values['source_url'] = $source_url;
-		$values['target_url'] = $target_url;
-
-		// Details about the user that browsed here
-		$values['ip_address'] = $this->CI->input->ip_address();
-		$values['user_agent'] = $this->CI->input->user_agent();
-
-		// Including their login id if they are already logged in
-		if($this->CI->user->loggedIn())
+		if($this->collectStats)
 		{
-			$values['login_identifier'] = $this->CI->user->identifier();
+			$values = array();
+
+			// The server time of the request
+			$values['timestamp'] = time();
+
+			// The source URI (what page was the link originally on)
+			$values['source_url'] = $source_url;
+			$values['target_url'] = $target_url;
+
+			// Details about the user that browsed here
+			$values['ip_address'] = $this->CI->input->ip_address();
+			$values['user_agent'] = $this->CI->input->user_agent();
+
+			// Including their login id if they are already logged in
+			if($this->CI->user->loggedIn())
+			{
+				$values['login_identifier'] = $this->CI->user->identifier();
+			}
+
+			// Optionally, a note for whatever use...
+			if ($note) { $values['note'] = $note; }
+
+			$this->db->insert('click_stats', $values);
 		}
-
-		// Optionally, a note for whatever use...
-		if ($note) { $values['note'] = $note; }
-
-		$this->db->insert('click_stats', $values);
 	}
 
 
@@ -94,50 +101,53 @@ class Stats {
 	 */
 	public function registerSearchTerm($search_term, $num_found, $note=null)
 	{
-		$values = array();
+		if($this->collectStats)
+		{
+			$values = array();
 
-		// The server time of the request
-		$values['timestamp'] = time();
+			// The server time of the request
+			$values['timestamp'] = time();
 
-		// The term which was searched for
-		$values['term'] = $search_term;
-		$values['num_found'] = $num_found;
+			// The term which was searched for
+			$values['term'] = $search_term;
+			$values['num_found'] = $num_found;
 
-		// Details about the user that browsed here
-		$values['ip_address'] = $this->CI->input->ip_address();
-		$values['user_agent'] = $this->CI->input->user_agent();
+			// Details about the user that browsed here
+			$values['ip_address'] = $this->CI->input->ip_address();
+			$values['user_agent'] = $this->CI->input->user_agent();
 
-		// Including their login id if they are already logged in
-		if($this->CI->user->loggedIn()){
-			$values['login_identifier'] = $this->CI->user->identifier();
-		}
+			// Including their login id if they are already logged in
+			if($this->CI->user->loggedIn()){
+				$values['login_identifier'] = $this->CI->user->identifier();
+			}
 
-		// Optionally, a note for whatever use...
-		if ($note) { $values['note'] = $note; }
+			// Optionally, a note for whatever use...
+			if ($note) { $values['note'] = $note; }
 
-		$this->db->insert('search_terms', $values);
+			$this->db->insert('search_terms', $values);
 
-		//increment or update search_occurence and update the ranking
-		
-		$existing_terms = $this->db->select('*')->where('term', $search_term)->limit(1)->get('search_occurence');
-		if($existing_terms->num_rows() == 0){
-			//there is none, add a new one
-			$new_term['term'] = $search_term;
-			$new_term['occurence'] = 1;
-			$new_term['num_found'] = $num_found;
-			$new_term['ranking'] = $this->calculate_ranking(1, $num_found);
-			$this->db->insert('search_occurence', $new_term);
-		}else{
-			//there is already 1
-			foreach($existing_terms->result() as $term){
-				$new_occurence = $term->occurence + 1;
-				$data = array(
-					'occurence' => $new_occurence,
-					'num_found' => $num_found,
-					'ranking' => $this->calculate_ranking($new_occurence, $num_found)
-				);
-				$this->db->where('term', $search_term);
-				$this->db->update('search_occurence', $data);
+			//increment or update search_occurence and update the ranking
+			
+			$existing_terms = $this->db->select('*')->where('term', $search_term)->limit(1)->get('search_occurence');
+			if($existing_terms->num_rows() == 0){
+				//there is none, add a new one
+				$new_term['term'] = $search_term;
+				$new_term['occurence'] = 1;
+				$new_term['num_found'] = $num_found;
+				$new_term['ranking'] = $this->calculate_ranking(1, $num_found);
+				$this->db->insert('search_occurence', $new_term);
+			}else{
+				//there is already 1
+				foreach($existing_terms->result() as $term){
+					$new_occurence = $term->occurence + 1;
+					$data = array(
+						'occurence' => $new_occurence,
+						'num_found' => $num_found,
+						'ranking' => $this->calculate_ranking($new_occurence, $num_found)
+					);
+					$this->db->where('term', $search_term);
+					$this->db->update('search_occurence', $data);
+				}
 			}
 		}
 	}
@@ -149,7 +159,12 @@ class Stats {
 	 * @return int ranking
 	 */
 	private function calculate_ranking($occurence, $num_found){
-		return $occurence * log($num_found);
+		if($this->collectStats)
+		{
+			return $occurence * log($num_found);
+		}
+		else
+			return 1;
 	}
 
 	/**
@@ -160,7 +175,8 @@ class Stats {
 	public function getSearchSuggestion($like)
 	{
 		$result = array();
-		if($like){
+		if($this->collectStats && $like)
+		{
 			$this->db->select('term')->order_by('ranking', 'desc')->limit(5)->like('term', $like)->where('ranking >', 0);
 			$matches = $this->db->get('search_occurence');
 			foreach($matches->result() as $match){
@@ -178,19 +194,22 @@ class Stats {
 	 */
 	public function registerSearchStats($search_term, $occurence)
 	{
-		$values = array();
+		if($this->collectStats)
+		{
+			$values = array();
 
-		// The server time of the request
-		$values['timestamp'] = time();
+			// The server time of the request
+			$values['timestamp'] = time();
 
-		// The term which was searched for
-		$values['search_term'] = $search_term;
+			// The term which was searched for
+			$values['search_term'] = $search_term;
 
-		// The number of objects returned from the search
-		$values['occurrence'] = $occurence;
-	
+			// The number of objects returned from the search
+			$values['occurrence'] = $occurence;
+		
 
-		$this->db->insert('search_result_counts', $values);
+			$this->db->insert('search_result_counts', $values);
+		}
 	}
 
 	/**
@@ -202,47 +221,50 @@ class Stats {
 	 */
 	public function updateSearchTermOccurence($since = null)
 	{
-		if ($since)
+		if($this->collectStats)
 		{
-			// This mode is "additive" for all entries since time $since
-			// Select matching terms
-			$this->db->select('term, COUNT(*) AS occurence', FALSE)->where('timestamp >=', $since)->group_by('term');
-			$matches = $this->db->get('search_terms');
-
-			foreach($matches->result_array() AS $match)
+			if ($since)
 			{
-				$this->db->where('term', $match['term'])->select('occurence');
-				$aggregate_count = $this->db->get('search_occurence');
+				// This mode is "additive" for all entries since time $since
+				// Select matching terms
+				$this->db->select('term, COUNT(*) AS occurence', FALSE)->where('timestamp >=', $since)->group_by('term');
+				$matches = $this->db->get('search_terms');
 
-				// This function will always return a result (count = 0  if term does not exist)
-				$aggregate_count = array_pop($aggregate_count->result_array());
+				foreach($matches->result_array() AS $match)
+				{
+					$this->db->where('term', $match['term'])->select('occurence');
+					$aggregate_count = $this->db->get('search_occurence');
 
-				if ((int)$aggregate_count['occurence'] > 0)
-				{
-					// There already exists an occurence within the aggregate table
-					// so ADD to it
-					$this->db->where('term',$match['term']);
-					$this->db->update('search_occurence', array('occurence' => ((int)$aggregate_count['occurence'] + (int)$match['occurence'])));
+					// This function will always return a result (count = 0  if term does not exist)
+					$aggregate_count = array_pop($aggregate_count->result_array());
+
+					if ((int)$aggregate_count['occurence'] > 0)
+					{
+						// There already exists an occurence within the aggregate table
+						// so ADD to it
+						$this->db->where('term',$match['term']);
+						$this->db->update('search_occurence', array('occurence' => ((int)$aggregate_count['occurence'] + (int)$match['occurence'])));
+					}
+					else
+					{
+						// Otherwise just add it 
+						$this->db->insert('search_occurence', array('term'=> $match['term'], 'occurence' => $match['occurence']));
+					}
 				}
-				else
-				{
-					// Otherwise just add it 
-					$this->db->insert('search_occurence', array('term'=> $match['term'], 'occurence' => $match['occurence']));
-				}
+
 			}
-
-		}
-		else
-		{
-			// This mode is like a "Resync" (it replaces all existing aggregates)
-			$this->db->truncate('search_occurence');
-
-			$this->db->select('term, COUNT(*) AS occurence', FALSE)->group_by('term');
-			$matches = $this->db->get('search_terms');
-
-			foreach($matches->result_array() AS $match)
+			else
 			{
-				$this->db->insert('search_occurence', array('term'=>$match['term'], 'occurence'=>$match['occurence']));
+				// This mode is like a "Resync" (it replaces all existing aggregates)
+				$this->db->truncate('search_occurence');
+
+				$this->db->select('term, COUNT(*) AS occurence', FALSE)->group_by('term');
+				$matches = $this->db->get('search_terms');
+
+				foreach($matches->result_array() AS $match)
+				{
+					$this->db->insert('search_occurence', array('term'=>$match['term'], 'occurence'=>$match['occurence']));
+				}
 			}
 		}
 	}
@@ -251,9 +273,12 @@ class Stats {
 	public function Stats()
 	{
 		$this->CI =& get_instance();
-
+		$this->collectStats = $this->CI->config->item('collect_stats');
 		// setup the DB connection
-		$this->db = $this->CI->load->database('portal', TRUE);
+		if($this->collectStats)
+		{
+			$this->db = $this->CI->load->database('portal', TRUE);
+		}
 	}
